+++ adapters/perception/src/errors.mjs
@@ -1,0 +1,7 @@
+export class PerceptionError extends Error {
+  constructor(code, message) {
+    super(message);
+    this.name = "PerceptionError";
+    this.code = code;
+  }
+}